requirements.txt
python-dotenv>=1.0.0
requests>=2.32.5
tqdm>=4.67.1